+++ internal/pkg/runtime/launcher/oci/mounts_linux.go
@@ -58,6 +58,11 @@ func (l *Launcher) getMounts() ([]specs.Mount, error) {
 
 // addTmpMounts adds tmpfs mounts for /tmp and /var/tmp in the container.
 func (l *Launcher) addTmpMounts(mounts *[]specs.Mount) {
+	if !l.singularityConf.MountTmp {
+		sylog.Debugf("Skipping mount of /tmp due to singularity.conf")
+		return
+	}
+
 	*mounts = append(*mounts,
 
 		specs.Mount{
@@ -139,6 +144,11 @@ func (l *Launcher) addDevMounts(mounts *[]specs.Mount) error {
 
 // addProcMount adds the /proc tree in the container.
 func (l *Launcher) addProcMount(mounts *[]specs.Mount) {
+	if !l.singularityConf.MountProc {
+		sylog.Debugf("Skipping mount of /proc due to singularity.conf")
+		return
+	}
+
 	*mounts = append(*mounts,
 		specs.Mount{
 			Source:      "proc",
@@ -149,6 +159,11 @@ func (l *Launcher) addProcMount(mounts *[]specs.Mount) {
 
 // addSysMount adds the /sys tree in the container.
 func (l *Launcher) addSysMount(mounts *[]specs.Mount) {
+	if !l.singularityConf.MountSys {
+		sylog.Debugf("Skipping mount of /sys due to singularity.conf")
+		return
+	}
+
 	if os.Getuid() == 0 {
 		*mounts = append(*mounts,
 			specs.Mount{
@@ -172,6 +187,11 @@ func (l *Launcher) addSysMount(mounts *[]specs.Mount) {
 // emulating `--compat` / `--containall`, so the user must specifically bind in
 // their home directory from the host for it to be available.
 func (l *Launcher) addHomeMount(mounts *[]specs.Mount) error {
+	if !l.singularityConf.MountHome {
+		sylog.Debugf("Skipping mount of $HOME due to singularity.conf")
+		return nil
+	}
+
 	// Get the host user's data
 	pw, err := user.CurrentOriginal()
 	if err != nil {