+++ CHANGELOG.md
@@ -35,6 +35,10 @@
   (e.g. ssh login session scope).
 - Pass host's `TERM` environment variable to container in OCI mode. Can be
   overridden by setting `SINGULARITYENV_TERM` on host.
+- Honour `config passwd` and `config group` directives from `singularity.conf`
+  in `--oci` mode.
+- Honour `mount proc` / `mount sys` / `mount tmp` / `mount home` directives from
+  `singularity.conf` in `--oci` mode.
 
 ## 3.11.1 \[2023-03-14\]